raid to module
