+++ frontend/src/interfaces/State.ts
@@ -56,8 +56,6 @@ export interface IRaidState {
   durabilityCost: string;
   xpReward: string;
   accountPower: string;
-
-  //isOwnedCharacterRaidingById: Record<number, boolean>; // ?
 }
 export interface IPartnerProject {
   id: string;
@@ -186,8 +184,6 @@ export interface IState {
 
   staking: Record<StakeType | NftStakeType, IStakeState>;
   stakeOverviews: Record<StakeType | NftStakeType, IStakeOverviewState>;
-
-  raid: IRaidState;
 
   waxBridgeWithdrawableBnb: string;
   waxBridgeRemainingWithdrawableBnbDuringPeriod: string;
